feat: refactor FounderFormItemTable to use RegionAutocomplete component
